// src/API.ts
export enum AssemblyType {
  RTA = "RTA",
  ASSEMBLED = "ASSEMBLED",
}
// ...
export type CreateCabinetProductMutationVariables = {
  input: CreateCabinetProductInput,
  condition?: ModelCabinetProductConditionInput | null,
};

export type CreateCabinetProductMutation = {
  createCabinetProduct?:  {
    __typename: "CabinetProduct",
    id: string,
    wSKU: string,
    vSKU: string,
    brand: string,
    doorStyle: string,
    cabinetType: CabinetType,
    cabinetSubType?: AssemblyType | null,
    productType: ProductType,
    productSubType?: ProductSubType | null,
    retailPrice?: number | null,
    discountScalingOn?: number | null,
    discountScalingOff?: number | null,
    listPrice?: number | null,
    assemblyMarkupFactor?: number | null,
    assemblyFee?: number | null,
    costFactor?: number | null,
    unitCost?: number | null,
    doesVendorAssemble?: boolean | null,
    vendorAssemblyCost?: number | null,
    waverlyAssemblyCost?: number | null,
    height?: number | null,
    width?: number | null,
    weight?: number | null,
    line?: CabinetLine | null,
    finish?: FinishType | null,
    finishColor?: string | null,
    overlay?: OverlayType | null,
    boxConstruction?: BoxConstructionType | null,
    doors?: string | null,
    drawers?: string | null,
    shelves?: string | null,
    species?: string | null,
    shippingClass?: AssemblyType | null,
    fullHeight?: boolean | null,
    pantry?: boolean | null,
    oven?: boolean | null,
    corner?: boolean | null,
    blindCabinet?: boolean | null,
    flipUpDoor?: boolean | null,
    pullout?: boolean | null,
    sink?: boolean | null,
    numberOfSinks?: number | null,
    microwave?: boolean | null,
    microWaveDrawer?: boolean | null,
    range?: boolean | null,
    rangeDrawer?: boolean | null,
    desk?: boolean | null,
    lineDrawingImagePath?: string | null,
    doorStyleImagePath?: string | null,
    tags?: string | null,
    stockItem?: boolean | null,
    availableForPickup?: boolean | null,
    publish?: boolean | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type UpdateCabinetProductMutationVariables = {
  input: UpdateCabinetProductInput,
  condition?: ModelCabinetProductConditionInput | null,
};

export type UpdateCabinetProductMutation = {
  updateCabinetProduct?:  {
    __typename: "CabinetProduct",
    id: string,
    wSKU: string,
    vSKU: string,
    brand: string,
    doorStyle: string,
    cabinetType: CabinetType,
    cabinetSubType?: AssemblyType | null,
    productType: ProductType,
    productSubType?: ProductSubType | null,
    retailPrice?: number | null,
    discountScalingOn?: number | null,
    discountScalingOff?: number | null,
    listPrice?: number | null,
    assemblyMarkupFactor?: number | null,
    assemblyFee?: number | null,
    costFactor?: number | null,
    unitCost?: number | null,
    doesVendorAssemble?: boolean | null,
    vendorAssemblyCost?: number | null,
    waverlyAssemblyCost?: number | null,
    height?: number | null,
    width?: number | null,
    weight?: number | null,
    line?: CabinetLine | null,
    finish?: FinishType | null,
    finishColor?: string | null,
    overlay?: OverlayType | null,
    boxConstruction?: BoxConstructionType | null,
    doors?: string | null,
    drawers?: string | null,
    shelves?: string | null,
    species?: string | null,
    shippingClass?: AssemblyType | null,
    fullHeight?: boolean | null,
    pantry?: boolean | null,
    oven?: boolean | null,
    corner?: boolean | null,
    blindCabinet?: boolean | null,
    flipUpDoor?: boolean | null,
    pullout?: boolean | null,
    sink?: boolean | null,
    numberOfSinks?: number | null,
    microwave?: boolean | null,
    microWaveDrawer?: boolean | null,
    range?: boolean | null,
    rangeDrawer?: boolean | null,
    desk?: boolean | null,
    lineDrawingImagePath?: string | null,
    doorStyleImagePath?: string | null,
    tags?: string | null,
    stockItem?: boolean | null,
    availableForPickup?: boolean | null,
    publish?: boolean | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type DeleteCabinetProductMutationVariables = {
  input: DeleteCabinetProductInput,
  condition?: ModelCabinetProductConditionInput | null,
};

export type DeleteCabinetProductMutation = {
  deleteCabinetProduct?:  {
    __typename: "CabinetProduct",
    id: string,
    wSKU: string,
    vSKU: string,
    brand: string,
    doorStyle: string,
    cabinetType: CabinetType,
    cabinetSubType?: AssemblyType | null,
    productType: ProductType,
    productSubType?: ProductSubType | null,
    retailPrice?: number | null,
    discountScalingOn?: number | null,
    discountScalingOff?: number | null,
    listPrice?: number | null,
    assemblyMarkupFactor?: number | null,
    assemblyFee?: number | null,
    costFactor?: number | null,
    unitCost?: number | null,
    doesVendorAssemble?: boolean | null,
    vendorAssemblyCost?: number | null,
    waverlyAssemblyCost?: number | null,
    height?: number | null,
    width?: number | null,
    weight?: number | null,
    line?: CabinetLine | null,
    finish?: FinishType | null,
    finishColor?: string | null,
    overlay?: OverlayType | null,
    boxConstruction?: BoxConstructionType | null,
    doors?: string | null,
    drawers?: string | null,
    shelves?: string | null,
    species?: string | null,
    shippingClass?: AssemblyType | null,
    fullHeight?: boolean | null,
    pantry?: boolean | null,
    oven?: boolean | null,
    corner?: boolean | null,
    blindCabinet?: boolean | null,
    flipUpDoor?: boolean | null,
    pullout?: boolean | null,
    sink?: boolean | null,
    numberOfSinks?: number | null,
    microwave?: boolean | null,
    microWaveDrawer?: boolean | null,
    range?: boolean | null,
    rangeDrawer?: boolean | null,
    desk?: boolean | null,
    lineDrawingImagePath?: string | null,
    doorStyleImagePath?: string | null,
    tags?: string | null,
    stockItem?: boolean | null,
    availableForPickup?: boolean | null,
    publish?: boolean | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type CreateSalesOrderMutationVariables = {
  input: CreateSalesOrderInput,
  condition?: ModelSalesOrderConditionInput | null,
};

export type CreateSalesOrderMutation = {
  createSalesOrder?:  {
    __typename: "SalesOrder",
    id: string,
    name: string,
    email: string,
    phoneNumber?: string | null,
    source?: OrderSource | null,
    vendor?: string | null,
    vendorNumber?: string | null,
    doorStyles?: string | null,
    PurchaseOrderId: string,
    customerRef?: string | null,
    salesRepRef?: string | null,
    transactionID?: string | null,
    transactionNumber?: string | null,
    transactionDate?: string | null,
    paymentStatus?: PaymentStatus | null,
    orderStatus?: OrderStatus | null,
    fulfillmentStatus?: FulfillmentStatus | null,
    shippingStatus?: ShippingStatus | null,
    billing?:  {
      __typename: "Address",
      contact?: string | null,
      line1?: string | null,
      line2?: string | null,
      city?: string | null,
      state?: string | null,
      postalCode?: string | null,
    } | null,
    shipping?:  {
      __typename: "Address",
      contact?: string | null,
      line1?: string | null,
      line2?: string | null,
      city?: string | null,
      state?: string | null,
      postalCode?: string | null,
    } | null,
    dueDate?: string | null,
    shipDate?: string | null,
    estimatedDeliveryDate?: string | null,
    subtotal?: number | null,
    taxAmount?: number | null,
    shippingAmount?: number | null,
    discountAmount?: number | null,
    totalAmount?: number | null,
    paymentMethod?: string | null,
    paymentGateway?: string | null,
    currency?: CurrencyCode | null,
    shippingMethod?: string | null,
    trackingNumber?: string | null,
    trackingURL?: string | null,
    carrier?: string | null,
    notes?: string | null,
    customerNotes?: string | null,
    metadata?: string | null,
    items?:  {
      __typename: "ModelSalesOrderItemsConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    payloads?:  {
      __typename: "ModelIntegrationPayloadConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type UpdateSalesOrderMutationVariables = {
  input: UpdateSalesOrderInput,
  condition?: ModelSalesOrderConditionInput | null,
};

export type UpdateSalesOrderMutation = {
  updateSalesOrder?:  {
    __typename: "SalesOrder",
    id: string,
    name: string,
    email: string,
    phoneNumber?: string | null,
    source?: OrderSource | null,
    vendor?: string | null,
    vendorNumber?: string | null,
    doorStyles?: string | null,
    PurchaseOrderId: string,
    customerRef?: string | null,
    salesRepRef?: string | null,
    transactionID?: string | null,
    transactionNumber?: string | null,
    transactionDate?: string | null,
    paymentStatus?: PaymentStatus | null,
    orderStatus?: OrderStatus | null,
    fulfillmentStatus?: FulfillmentStatus | null,
    shippingStatus?: ShippingStatus | null,
    billing?:  {
      __typename: "Address",
      contact?: string | null,
      line1?: string | null,
      line2?: string | null,
      city?: string | null,
      state?: string | null,
      postalCode?: string | null,
    } | null,
    shipping?:  {
      __typename: "Address",
      contact?: string | null,
      line1?: string | null,
      line2?: string | null,
      city?: string | null,
      state?: string | null,
      postalCode?: string | null,
    } | null,
    dueDate?: string | null,
    shipDate?: string | null,
    estimatedDeliveryDate?: string | null,
    subtotal?: number | null,
    taxAmount?: number | null,
    shippingAmount?: number | null,
    discountAmount?: number | null,
    totalAmount?: number | null,
    paymentMethod?: string | null,
    paymentGateway?: string | null,
    currency?: CurrencyCode | null,
    shippingMethod?: string | null,
    trackingNumber?: string | null,
    trackingURL?: string | null,
    carrier?: string | null,
    notes?: string | null,
    customerNotes?: string | null,
    metadata?: string | null,
    items?:  {
      __typename: "ModelSalesOrderItemsConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    payloads?:  {
      __typename: "ModelIntegrationPayloadConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type DeleteSalesOrderMutationVariables = {
  input: DeleteSalesOrderInput,
  condition?: ModelSalesOrderConditionInput | null,
};
// ...
export type CreateSalesOrderItemsMutationVariables = {
  input: CreateSalesOrderItemsInput,
  condition?: ModelSalesOrderItemsConditionInput | null,
};

export type CreateSalesOrderItemsMutation = {
  createSalesOrderItems?:  {
    __typename: "SalesOrderItems",
    id: string,
    txnLineID?: string | null,
    PurchaseOrderId?: string | null,
    itemListID?: string | null,
    itemFullName?: string | null,
    description?: string | null,
    assembly?: string | null,
    quantity?: number | null,
    rate?: number | null,
    total?: number | null,
    invoiced?: number | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    salesOrderItemsId?: string | null,
  } | null,
};